requirements.txt
streamlit>=1.32.0
langgraph>=0.2.50
langchain-core>=0.3.0
langchain-anthropic>=0.2.0
anthropic>=0.40.0
tavily-python>=0.5.0
edgartools>=2.0.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.20.0
reportlab>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
